third_party/blink/web_tests/html/tabular_data/table_exceptions-expected.txt

This test should trigger exceptions on HTMLTableElement, and verify that the messages are reasonably helpful.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS t.caption = document.body threw exception TypeError: Failed to set the 'caption' property on 'HTMLTableElement': Failed to convert value to 'HTMLTableCaptionElement'..
PASS t.tHead = document.body threw exception TypeError: Failed to set the 'tHead' property on 'HTMLTableElement': Failed to convert value to 'HTMLTableSectionElement'..
PASS t.createTFoot() is t.tFoot
PASS t.tFoot is non-null.
PASS t.tFoot = null did not throw exception.
PASS t.tFoot is null
PASS t.tFoot = document.body threw exception TypeError: Failed to set the 'tFoot' property on 'HTMLTableElement': Failed to convert value to 'HTMLTableSectionElement'..
PASS t.insertRow(-2) threw exception IndexSizeError: Failed to execute 'insertRow' on 'HTMLTableElement': The index provided (-2) is less than -1..
PASS t.insertRow(1) threw exception IndexSizeError: Failed to execute 'insertRow' on 'HTMLTableElement': The index provided (1) is greater than the number of rows in the table (0)..
PASS t.deleteRow(-2) threw exception IndexSizeError: Failed to execute 'deleteRow' on 'HTMLTableElement': The index provided (-2) is less than -1..
PASS t.deleteRow(2) threw exception IndexSizeError: Failed to execute 'deleteRow' on 'HTMLTableElement': The index provided (2) is greater than the number of rows in the table (1)..
PASS t.deleteRow() threw exception TypeError: Failed to execute 'deleteRow' on 'HTMLTableElement': 1 argument required, but only 0 present..
PASS successfullyParsed is true

TEST COMPLETE
